+++ CHANGELOG.md
@@ -5,3 +5,5 @@ All notable changes to this project will be documented in this file.
 ## v1.0.7
 
 - UMD, improved gulpfile.js. Refactor test folder.
+
+- riccioStore, getInstance(), isInitialized(). Deprecated getStore().